new state + shr_shl + stop
