+++ zkevm-circuits/src/evm_circuit/witness.rs
@@ -17,6 +17,7 @@ use eth_types::evm_types::OpcodeId;
 use eth_types::{Address, Field, ToLittleEndian, ToScalar, ToWord, Word};
 use eth_types::{ToAddress, U256};
 use halo2_proofs::arithmetic::{BaseExt, FieldExt};
+use halo2_proofs::pairing::bn256::Fr as Fp;
 use itertools::Itertools;
 use sha3::{Digest, Keccak256};
 use std::{collections::HashMap, convert::TryInto, iter};
